Remove int cast from PercentFull method

+++ Content.Server/Chemistry/EntitySystems/SolutionContainerSystem.Capabilities.cs
@@ -113,12 +113,12 @@ public sealed partial class SolutionContainerSystem
             : solution.CurrentVolume;
     }
 
-    public int PercentFull(EntityUid uid)
+    public float PercentFull(EntityUid uid)
     {
         if (!TryGetDrainableSolution(uid, out var solution) || solution.MaxVolume.Equals(FixedPoint2.Zero))
             return 0;
 
-        return (int)((solution.CurrentVolume.Float() / solution.MaxVolume.Float()) * 100);
+        return ((solution.CurrentVolume.Float() / solution.MaxVolume.Float()) * 100);
     }
 
     public bool TryGetFitsInDispenser(EntityUid owner,
